requirements.txt
functions-framework==3.*
line-bot-sdk
google-generativeai